Update original classes to base classes
