+++ src/app/@shared/class/player.class.ts
@@ -1,4 +1,4 @@
-import { EspnClientEvent, EspnClientPlayer } from '@app/espn/mlb/interface';
+import { EspnClientPlayer } from '@app/espn/mlb/interface';
 
 /**
  * Base/Parent Class for all ESPN _fantasy_ players
@@ -11,5 +11,32 @@ export class Player {
 
   constructor(player: EspnClientPlayer) {
     this._player = player;
+  }
+
+  get id() {
+    return this._player.playerId;
+  }
+
+  get name() {
+    return this._player.playerPoolEntry.player.fullName;
+  }
+
+  get isInjured() {
+    return this._player.playerPoolEntry.player.injured;
+  }
+
+  get injuryStatus() {
+    return this._player.playerPoolEntry.player.injuryStatus;
+  }
+
+  get playerRatings() {
+    return this._player.playerPoolEntry.ratings;
+  }
+
+  get playerOwnership() {
+    return {
+      change: this._player.playerPoolEntry.player.ownership.percentChange,
+      percentOwned: this._player.playerPoolEntry.player.ownership.percentOwned,
+    };
   }
 }
