+++ src/app/espn/mlb/class/baseballPlayer.class.ts
@@ -4,6 +4,7 @@ import { MLB_LINEUP } from '../consts/lineup.const';
 import { MLB_POSITION } from '../consts/position.const';
 import { MLB_TEAM } from '../consts/team.const';
 import { weights2021 } from '../mlb.const';
+import { Player } from '@app/@shared/class';
 
 /**
  * This is a description of the BaseballPlayer constructor function.
@@ -11,32 +12,11 @@ import { weights2021 } from '../mlb.const';
  * @class
  * @classdesc This is a description of the BaseballPlayer class.
  */
-export class BaseballPlayer {
+export class BaseballPlayer extends Player {
   readonly weights21 = weights2021;
 
-  private _player: EspnClientPlayer;
   private _startingStatus: EspnClientGameStatus;
   private _isStarting = true;
-
-  constructor(player: EspnClientPlayer) {
-    this._player = player;
-  }
-
-  get id() {
-    return this._player.playerId;
-  }
-
-  get name() {
-    return this._player.playerPoolEntry.player.fullName;
-  }
-
-  get isInjured() {
-    return this._player.playerPoolEntry.player.injured;
-  }
-
-  get injuryStatus() {
-    return this._player.playerPoolEntry.player.injuryStatus;
-  }
 
   get playerImg() {
     return `https://a.espncdn.com/combiner/i?img=/i/headshots/mlb/players/full/${this._player.playerId}.png&w=96&h=70&cb=1`;
@@ -90,17 +70,6 @@ export class BaseballPlayer {
 
   get startingStatus() {
     return this._startingStatus;
-  }
-
-  get playerRatings() {
-    return this._player.playerPoolEntry.ratings;
-  }
-
-  get playerOwnership() {
-    return {
-      change: this._player.playerPoolEntry.player.ownership.percentChange,
-      percentOwned: this._player.playerPoolEntry.player.ownership.percentOwned,
-    };
   }
 
   get isPitcher() {
